Day-3/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
numpy==1.26.4
scipy==1.13.0
opencv-python==4.10.0.84
Pillow==10.3.0
pymupdf==1.24.0
pytesseract==0.3.10
easyocr==1.7.2
paddlepaddle==3.0.0
paddleocr==2.8.1
torch==2.2.0
torchvision==0.17.0
scikit-image==0.24.0